[EI-999] [web] refactor: Add Chip component for prompt presets and update PromptPreset component
